Compile error of probably new ver of tsc
Apparently, in the previous version of typescript, the `entry` attribute
in the webpack options was ignored, and apparently, it is no longer the
case. Apparently *(also)*, the value that were in our config files was
wrong (because `__dirname` was wrong) and this was breaking everything.

This commit removes the `entry` attribute of all config files (the entry
point is known by `tsc` thanks to the `rootDir` attribute of the
`tsconfig` file).